e2e_demo/mcp-app/requirements.txt
# General dependencies
PyYAML>=6.0
pydantic>=2.0.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
requests>=2.31.0

# OpenAI and Anthropic client libraries
openai>=1.5.0
anthropic>=0.5.0

# For MCP Agent
git+https://github.com/anthropics/mcp-agent.git